*FIXME* rails-mode and rspec-mode

+++ install.sh
@@ -9,19 +9,39 @@ usage()
     exit 1
 }
 
-rm -rf *.elc conf/*.elc
-touch myinfo.el
-echo "installing el-get package...."
-wget --no-check-certificate https://github.com/dimitri/el-get/raw/master/el-get-install.el
-emacs --script el-get-install.el
-rm el-get-install.el
-echo "installing el-get package....done"
-echo "installing other packages...."
-for setup_file in setup/*.el
-do
-    emacs -batch -l el-get/el-get/el-get.el -l elpa/package.el -l $setup_file 
-done
-echo "installing other packages....done"
+if [ $# -eq 1 ]
+then
+    setup_file="setup/el-get-$1.el"
+    if [ -f $setup_file ]
+    then
+	emacs -batch -l el-get/el-get/el-get.el -l elpa/package.el -l $setup_file 
+    else
+	setup_file=`grep -r $1 setup | line | cut -d ':' -f 1`
+	if [ -f $setup_file ]
+	then
+	emacs -batch -l el-get/el-get/el-get.el -l elpa/package.el -l $setup_file
+	fi
+    fi
+    
+fi
+
+
+if [ $# -eq 0 ]
+then
+    rm -rf *.elc conf/*.elc
+    touch myinfo.el
+    echo "installing el-get package...."
+    wget --no-check-certificate https://github.com/dimitri/el-get/raw/master/el-get-install.el
+    emacs --script el-get-install.el
+    rm el-get-install.el
+    echo "installing el-get package....done"
+    echo "installing other packages...."
+    for setup_file in setup/*.el
+    do
+	emacs -batch -l el-get/el-get/el-get.el -l elpa/package.el -l $setup_file 
+    done
+    echo "installing other packages....done"
+fi
 # for directory in ~/.emacs.d/*; do
 #     if [ -d $directory ]; then
 # 	LOAD_PATH="$LOAD_PATH -L $directory"
